[imp] qt::CsvTableView: lots of new distribution options in "Generate content"
-Almost all distributions in include file <random> are now available: 6 integer-valued and 11 real-valued distributions.
-Options are not visible user; to be improved later.
-Also done some tweaks to rand/distributionHelpers.hpp to silence warnings.

+++ dfg/rand/distributionHelpers.hpp
@@ -157,8 +157,8 @@ DFG_ROOT_NS_BEGIN{ DFG_SUB_NS(rand) {
             { \
                 return canConstruct_##CLASS_NAME<T>(args.second); \
             } \
-            static bool isAcceptableParamCount(const ptrdiff_t val) { return val < 32 && ((DFG_DETAIL_NS::ACCEPTED_ARG_COUNT & (1 << val)) != 0); } \
-            static const int MaxArgCount = DFG_DETAIL_NS::MaxArgCount<DFG_DETAIL_NS::ACCEPTED_ARG_COUNT>::value; \
+            static bool isAcceptableParamCount(const ptrdiff_t val) { return val < 32 && ((DFG_DETAIL_NS::ACCEPTED_ARG_COUNT & (ptrdiff_t(1) << val)) != 0); } \
+            static const ptrdiff_t MaxArgCount = DFG_DETAIL_NS::MaxArgCount<DFG_DETAIL_NS::ACCEPTED_ARG_COUNT>::value; \
             static std::pair<bool, std::tuple<__VA_ARGS__>> makeUninitializedParams() { return std::make_pair(false, std::tuple<__VA_ARGS__>()); } \
         }
